Seed the pseudorandom number generator only once
“Seeding” (= initializing) the pseudorandom number generator is
essentially a one-time, setup operation.

It’s a waste of resources to call srand() every time a random number
is needed. Doing so doesn’t increase the “randomness quality”; that
could even lower it. If randomness quality was a concern, using a
better PRNG would be the solution.

So call srand() only once, at the beginning of main().