Inline the only call to rand()
The function `randomnumber(int n)` was only called once, and given
the previous commit only contained a single line (`rand() % n`).
So it sounds better to just “inline” the call to `rand()` directly.

Also, the logic previously was overly complicated; a random number from
1 to 4 was selected, and then 1 was substracted from this number. Taking
a random number directly in the [0..3] range is simpler.